Exception handling -delete 2

+++ src/main/java/com/educamais/course/services/UserService.java
@@ -5,7 +5,6 @@ import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.educamais.course.entities.User;
@@ -33,12 +32,11 @@ public class UserService {
 	}
 	
 	public void delete(Long id) {
+		if(!repository.existsById(id)) throw new ResourceNotFoundException(id);
 		try {
 			repository.deleteById(id);
 		}
-		catch(EmptyResultDataAccessException e) {
-			throw new ResourceNotFoundException(id);
-		}
+		
 		catch(DataIntegrityViolationException e) {
 			throw new DatabaseException(e.getMessage());
 		}
